@@ -11,29 +11,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "~8.2.14",
     "@angular/http": "^7.1.3",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@fortawesome/fontawesome-free": "^5.12.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.5",
     "@nguniversal/express-engine": "^7.0.2",
+    "@types/chart.js": "^2.9.8",
+    "angular-bootstrap-md": "^8.8.1",
     "angular-in-memory-web-api": "^0.7.0",
+    "animate.css": "^3.7.2",
+    "bootstrap": "^4.4.1",
+    "chart.js": "^2.5.0",
     "core-js": "^2.5.4",
+    "hammerjs": "^2.0.8",
     "ngx-cookie": "^4.1.2",
     "ngx-cookie-service": "^2.1.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.4",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.0",
-    "@angular/cli": "~7.1.0",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular-devkit/build-angular": "^0.803.22",
+    "@angular/cli": "~8.3.22",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -48,6 +58,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3"
   }
 }
